server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
camelot-py==0.9.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
cssselect==1.2.0
diskcache==5.6.1
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.100.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
google==3.0.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.31.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
openai==0.27.8
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
pdfquery==0.4.3
pdftables-api @ git+https://github.com/pdftables/python-pdftables-api.git@0abf88cc499ed6c001fb641791f84cdc8eafbc58
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyMuPDF==1.23.0
PyMuPDFb==1.23.0
PyPDF2==3.0.1
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
replicate==0.9.0
requests==2.31.0
roman==4.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
tabula-py==2.7.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
